refactor: consolidate template imports into a central index file
